Create the first version of personal .dotfiles
The proper machine setup was done by
github.com/jonathanbeber/machine.git, but this project was forked and
getting out of date. Also, the project was using Ansible, a tool that
I'm no longer insteressed in. Now I'm using Ubuntu machines 100% of my
time and want to keep things in a simpler approach.

This commit creates a first version which is resposible for
installing/updating some important software (in this project called
tools) and config them. I used symbolic links to keep configs under this
repo and in its own configuration dir at the same time.